bonus/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow
spacy
streamlit
pillow
